# application/services/custom_services/shiftypes/src/models/entitymodels/shiftypes.ts
export interface shiftypes 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   description: { type: String }
}